remove redundant functions
