+++ aitom/classify/deep/unsupervised/autoencoder/autoencoder_util.py
@@ -24,16 +24,7 @@ import os, sys, json, shutil, time, struct, array, pickle
 import copy
 import math
 
-def pickle_load(path):
-    with open(path,'rb') as f:
-        o = pickle.load(f,encoding='iso-8859-1')
-    return o
-
-def pickle_dump(o, path, protocol=-1):
-    with open(path, 'wb') as f:
-        pickle.dump(o, f, protocol=protocol)
-
-
+import aitom.image.vol.util as AIVU
 
 # def em2numpy(ve):
 #     assert ve.get_ndim() == 3           # only for 3D array
@@ -70,7 +61,7 @@ def peaks_to_subvolumes(v, peaks, subvol_size=32):
         d['vs'][uuid] = {}
         d['vs'][uuid]['center'] = p['x']
         d['vs'][uuid]['id'] = uuid
-        d['vs'][uuid]['v'] = cut_from_whole_map(v, p['x'], subvol_size)
+        d['vs'][uuid]['v'] = AIVU.cut_from_whole_map(v, p['x'], subvol_size)
     return d
     
     
@@ -81,22 +72,10 @@ def read_mrc_numpy_vol(path):
     return v
 
 
-# 3D gaussian filtering of a volume (v)
-# smoothing using scipy.ndimage.gaussian_filter
-def smooth(v, sigma):
-    assert  sigma > 0
-    return SN.gaussian_filter(input=v, sigma=sigma)
-
-
-# paste to a map given start and end coordinates
-def paste_to_whole_map__se(whole_map, vol, se):
-    whole_map[se[0,0]:se[0,1], se[1,0]:se[1,1], se[2,0]:se[2,1]] = vol
-
-
 def difference_of_gauss_function(size, sigma1, sigma2):
 
-    grid = grid_displacement_to_center(size)
-    dist_sq = grid_distance_sq_to_center(grid)
+    grid = AIVU.grid_displacement_to_center(size)
+    dist_sq = AIVU.grid_distance_sq_to_center(grid)
 
     del grid
 
@@ -104,46 +83,6 @@ def difference_of_gauss_function(size, sigma1, sigma2):
     dog -= (1 / ( (2 * N.pi)**(3.0/2.0)  * (sigma2**3)) ) * N.exp( - (dist_sq)  / (2.0 * (sigma2**2)))
 
     return dog
-
-
-def grid_displacement_to_center(size, mid_co=None):
-
-    size = N.array(size, dtype=N.float)
-    assert size.ndim == 1
-
-    if mid_co is None:            mid_co = (N.array(size) - 1) / 2          # IMPORTANT: following python convension, in index starts from 0 to size-1!!! So (siz-1)/2 is real symmetry center of the volume
- 
-    if size.size == 3:
-        # construct a gauss function whose center is at center of volume
-        grid = N.mgrid[0:size[0], 0:size[1], 0:size[2]]
-
-        for dim in range(3):
-            grid[dim, :, :, :] -= mid_co[dim]
-
-    elif size.size == 2:
-        # construct a gauss function whose center is at center of volume
-        grid = N.mgrid[0:size[0], 0:size[1]]
-
-        for dim in range(2):
-            grid[dim, :, :] -= mid_co[dim]
-
-    else:
-        assert False
-
-    return grid
-
-def grid_distance_sq_to_center(grid):
-    dist_sq = N.zeros(grid.shape[1:])
-    if grid.ndim == 4:
-        for dim in range(3):
-            dist_sq += N.squeeze(grid[dim, :, :, :]) ** 2
-    elif grid.ndim == 3:
-        for dim in range(2):
-            dist_sq += N.squeeze(grid[dim, :, :]) ** 2
-    else:
-        assert False
-
-    return dist_sq
 
 
 def put_mrc(mrc, path):
@@ -162,99 +101,6 @@ def write_data(data, path):
 
 
 
-
-def cut_from_whole_map(whole_map, c, siz):
-
-    se = subvolume_center_start_end(c, map_siz=whole_map.shape, subvol_siz=siz)
-    return          cut_from_whole_map__se(whole_map, se)
-
-
-# cut a map given start and end coordinates
-def cut_from_whole_map__se(whole_map, se):
-    if se is None:       return None
-    return          whole_map[se[0,0]:se[0,1], se[1,0]:se[1,1], se[2,0]:se[2,1]]
-
-
-
-# given a center c, get the relative start and end position of a subvolume with size subvol_siz
-def subvolume_center_start_end(c, map_siz, subvol_siz):
-    map_siz = N.array(map_siz)
-    subvol_siz = N.array(subvol_siz)
-
-    siz_h = N.ceil( subvol_siz / 2.0 )
-
-    start = c - siz_h;      start.astype(int)
-    end = start + subvol_siz;      end.astype(int)
-
-    if any(start < 0):  return None
-    if any(end >= map_siz):    return None
-
-    se = N.zeros( (3,2), dtype=N.int )
-    se[:,0] = start
-    se[:,1] = end
-
-    return se
-
-def cub_img(v, view_dir=2):
-    if view_dir == 0:
-        vt = N.transpose(v, [1,2,0])
-    elif view_dir == 1:
-        vt = N.transpose(v, [2,0,1])
-    elif view_dir == 2:
-        vt = v
-    
-    row_num = vt.shape[0] + 1
-    col_num = vt.shape[1] + 1
-    slide_num = vt.shape[2]
-    disp_len = int( N.ceil(N.sqrt(slide_num)) )
-    
-    slide_count = 0
-    im = N.zeros( (row_num*disp_len, col_num*disp_len) ) + float('nan')
-    for i in range(disp_len):
-        for j in range(disp_len):
-            im[(i*row_num) : ((i+1)*row_num-1),  (j*col_num) : ((j+1)*col_num-1)] = vt[:,:, slide_count]
-            slide_count += 1
-            
-            if (slide_count >= slide_num):
-                break
-            
-        
-        if (slide_count >= slide_num):
-            break
-   
-    
-    im_v = im[N.isfinite(im)]
-
-    if im_v.max() > im_v.min(): 
-        im = (im - im_v.min()) / (im_v.max() - im_v.min())
-
-    return {'im':im, 'vt':vt}
-
-def save_png(m, name, normalize=True, verbose=False):
-
-    if verbose:
-        print ('save_png()')
-        print ('unique values', sorted(set(m.flatten())))
-
-    m = N.array(m, dtype=N.float)
-
-    mv = m[N.isfinite(m)]
-    if normalize:
-        # normalize intensity to 0 to 1
-        if mv.max() - mv.min() > 0:
-            m = (m - mv.min()) / (mv.max() - mv.min())
-        else:
-            m = N.zeros(m.shape)
-    else:
-        assert mv.min() >= 0
-        assert mv.max() <= 1
-
-    m = N.ceil(m * 65534)
-    m = N.array(m, dtype=N.uint16)
-
-    import png          # in pypng package
-    png.from_array(m, 'L').save(name)
-
 def center_mass(v):
     assert  N.all(v >= 0)
     
@@ -272,8 +118,6 @@ def center_mass(v):
 
 
     return N.array(c)
-
-
 
 '''
 perform pose normalization through PCA, assume that the values in V is non-negative
@@ -360,7 +204,7 @@ def rotate_retrieve(v, tom, center = None, angle = None, rm=None, c1=None, c2=No
       
     siz = N.ceil((v.shape[0]+ 2*max(N.abs(loc_r)))*N.abs(rm.dot([1,1,1])).max()) #bounding box size
     if all((center - siz) > 0) and all((center + siz) < tom.shape): #test if the tomogram enclose the bounding box
-        vb = cut_from_whole_map(tom, center, siz)
+        vb = AIVU.cut_from_whole_map(tom, center, siz)
     
     else: 
         vb = N.full((int(siz),int(siz),int(siz)),default_val) #set non-enclosed regions as default value
@@ -417,7 +261,7 @@ def rotate_retrieve(v, tom, center = None, angle = None, rm=None, c1=None, c2=No
     
     c = N.array([(vbr.shape[0])/2]*3, dtype=int)
     siz = v.shape[0]
-    vr = cut_from_whole_map(vbr, c, siz) #crop the center image of original subtomogram size from the rotated bounding box
+    vr = AIVU.cut_from_whole_map(vbr, c, siz) #crop the center image of original subtomogram size from the rotated bounding box
     
     return vr
