+++ aitom/tomominer/dimension_reduction/util.py
@@ -15,7 +15,7 @@ import numpy as np
 from numpy.fft import fftn, ifftn, fftshift, ifftshift
 import aitom.tomominer.io.file as iv
 import aitom.tomominer.average.util as avgu
-import aitom.tomominer.image.vol.util as uv
+import aitom.image.vol.util as uv
 
 def neighbor_product(v):
     siz = list(v.shape)
